requirements.txt
bs4
amqp==2.2.2
Babel==2.5.1
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
defusedxml==0.5.0
Django==1.11.5
django-admin-list-filter-dropdown==1.0.1
django-admin-rangefilter==0.3.0
django-cors-headers==2.1.0
django-filter==1.0.4
django-filters==0.2.1
django-rest-swagger==2.1.2
djangorestframework==3.6.4
drfdocs==0.0.11
flower==0.9.2
grab==0.6.38
html5lib==1.0.1
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
kombu==4.1.0
lxml==3.8.0
MarkupSafe==1.0
numpy==1.13.3
openapi-codec==1.3.2
pandas==0.20.3
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
python-dateutil==2.6.1
pytils==0.3
pytrends==4.3.0
pytz==2017.3
raven==6.4.0
redis==2.10.6
requests==2.18.4
selection==0.0.13
selenium==3.6.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.10.0
tornado==4.5.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
user-agent==0.1.9
uWSGI==2.0.15
vaderSentiment==2.5
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
weblib==0.1.25
xlrd==1.1.0
XlsxWriter==1.0.0